Moved the global view to another component
I did this so it can be used in the 'new' component
