+++ app/assets/javascript/quotations/quotations.js
@@ -3345,6 +3345,18 @@ eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export *
 
 /***/ }),
 
+/***/ "./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/components/global_view.vue?vue&type=script&lang=js&":
+/*!******************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/lib??vue-loader-options!./viewsrc/modules/quotations/components/global_view.vue?vue&type=script&lang=js& ***!
+  \******************************************************************************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n\n /* harmony default export */ __webpack_exports__[\"default\"] = ({\n   props:{\n    section_valid: {\n      type: Boolean,\n      default: false\n    },\n    quotation_id:{\n      type:Number,\n      default: null\n    },\n    refresh_quotation:{\n      type: Boolean,\n      default: false\n    }\n  },\n\n   data() {\n    return {\n      brands:{\n        supranet: 'supranet',\n        siemon: 'siemon'\n      },\n      translations:{\n        show: I18n.t('quotations.show'),\n        types: I18n.t('quotations.api_types')\n      },\n      quotation: {\n        client: {},\n        quotation_products: [],\n        quotation_services: [],\n        products_totals: {},\n        services_totals:{}\n      },\n      quotation_products: [],\n      quotation_products_headers: [],\n      quotation_services_headers: []\n    }\n  },\n  \n  mounted: function(){\n    this.refreshQuotation();\n  },\n\n  methods: {\n\n    refreshQuotation(){\n      if(this.quotation_id){\n        this.getQuotation();\n        this.setQuotationServicesHeaders();\n      }\n    },\n    \n    getQuotation: function(){\n      this.http\n      .get(`/quotations/${this.quotation_id}`)\n      .then((response)=>{\n        if(response.successful){\n          this.quotation = response.data;\n          this.setQuotationProductsHeaders();\n        }else{\n          console.log(JSON.stringify(response));\n        }\n      }).catch((err)=>{\n        console.log(JSON.stringify(err));\n      });\n    },\n\n    setQuotationProductsHeaders: function(){\n      var headers = ['amount','material'];\n      var q_type = this.quotation.quotation_type;\n      //sets columns dependind on quotation_type, types are:\n      //t_comparative, t_supranet_only, t_siemon_only, t_simple\n      ['t_supranet_only','t_siemon_only'].forEach((type)=>{\n        if(q_type === 't_comparative' || q_type === type){\n          headers.push(`${type}_percent`);\n          //Adds four fields: unit price, total price, unit price plus percent and total price plus percent\n          ['','_with_percent'].forEach((suffix)=>{\n            headers = headers.concat([`${type}_price${suffix}`, `${type}_total${suffix}`]);\n          });\n        }\n      });\n      //simple type\n      if(q_type === 't_simple'){\n        headers = headers.concat(['brand','percent']);\n        ['','_with_percent'].forEach((suffix)=>{\n          headers = headers.concat([`${q_type}_price${suffix}`, `${q_type}_total${suffix}`]);\n        });\n      }\n      headers.forEach((element)=>{\n        this.quotation_products_headers.push({\n          key:element,\n          label:this.translations.show.headers[element],\n          sortable:true\n        });\n      });\n    },\n\n    setQuotationServicesHeaders: function(){\n      var headers = [\n        'amount',\n        'service',\n        'percent',\n        'price',\n        'total',\n        'price_with_percent',\n        'total_with_percent'\n      ].forEach((element)=>{\n        this.quotation_services_headers.push({\n          key:element,\n          label:this.translations.show.headers[element],\n          sortable:true\n        })\n      });\n    }\n  },\n\n  watch: {\n    refresh_quotation: function(){\n      if(this.refresh_quotation){\n        this.refresh_quotation = false;\n        this.refreshQuotation;\n      }\n    }\n  }\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/global_view.vue?./node_modules/vue-loader/lib??vue-loader-options");
+
+/***/ }),
+
 /***/ "./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/components/header.vue?vue&type=script&lang=js&":
 /*!*************************************************************************************************************************************!*\
   !*** ./node_modules/vue-loader/lib??vue-loader-options!./viewsrc/modules/quotations/components/header.vue?vue&type=script&lang=js& ***!
@@ -3401,7 +3413,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n\n/* harmony default export *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _components_header_vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./components/header.vue */ \"./viewsrc/modules/quotations/components/header.vue\");\n/* harmony import */ var _components_material_vue__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./components/material.vue */ \"./viewsrc/modules/quotations/components/material.vue\");\n/* harmony import */ var _components_conditions_vue__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/conditions.vue */ \"./viewsrc/modules/quotations/components/conditions.vue\");\n/* harmony import */ var _components_service_vue__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./components/service.vue */ \"./viewsrc/modules/quotations/components/service.vue\");\n\n\n\n\n\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n\n  components:{\n    'quotation-header' : _components_header_vue__WEBPACK_IMPORTED_MODULE_0__[\"default\"],\n    'quotation-materials' : _components_material_vue__WEBPACK_IMPORTED_MODULE_1__[\"default\"],\n    'quotation-service' : _components_service_vue__WEBPACK_IMPORTED_MODULE_3__[\"default\"],\n    'quotation-conditions': _components_conditions_vue__WEBPACK_IMPORTED_MODULE_2__[\"default\"]\n  },\n\n  data(){\n    return {\n      translations: I18n.t('quotations.new'),\n      tab_index: 0,\n      //these are only used to change the color of the tabs to green when the user finishes a section\n      completed:{\n        header:false,\n        materials:false,\n        services:false,\n        conditions:false,\n        global_view:false\n      },\n      quotation:{\n        id: NaN,\n      }\n    }\n  },\n\n  watch:{\n    'completed.header': function(){\n      if(this.completed.header){\n        this.tab_index++;\n      }\n    },\n    'completed.conditions': function(){\n      if(this.completed.conditions){\n        this.tab_index++;\n      }\n    },\n    'completed.services': function(){\n      if(this.completed.services){\n        this.tab_index++;\n      }\n    },\n    'completed.materials': function(){\n      if(this.completed.materials){\n        this.tab_index++;\n      }\n    }\n  }\n\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/new.vue?./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _components_header_vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./components/header.vue */ \"./viewsrc/modules/quotations/components/header.vue\");\n/* harmony import */ var _components_material_vue__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./components/material.vue */ \"./viewsrc/modules/quotations/components/material.vue\");\n/* harmony import */ var _components_conditions_vue__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/conditions.vue */ \"./viewsrc/modules/quotations/components/conditions.vue\");\n/* harmony import */ var _components_service_vue__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./components/service.vue */ \"./viewsrc/modules/quotations/components/service.vue\");\n\n\n\n\n\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n\n  components:{\n    'quotation-header' : _components_header_vue__WEBPACK_IMPORTED_MODULE_0__[\"default\"],\n    'quotation-materials' : _components_material_vue__WEBPACK_IMPORTED_MODULE_1__[\"default\"],\n    'quotation-service' : _components_service_vue__WEBPACK_IMPORTED_MODULE_3__[\"default\"],\n    'quotation-conditions': _components_conditions_vue__WEBPACK_IMPORTED_MODULE_2__[\"default\"]\n  },\n\n  data(){\n    return {\n      translations: I18n.t('quotations.new'),\n      tab_index: 0,\n      //these are only used to change the color of the tabs to green when the user finishes a section\n      completed:{\n        header:false,\n        materials:false,\n        services:false,\n        conditions:false,\n        global_view:false\n      },\n      quotation:{\n        id: null,\n      }\n    }\n  },\n\n  watch:{\n    'completed.header': function(){\n      if(this.completed.header){\n        this.tab_index++;\n      }\n    },\n    'completed.conditions': function(){\n      if(this.completed.conditions){\n        this.tab_index++;\n      }\n    },\n    'completed.services': function(){\n      if(this.completed.services){\n        this.tab_index++;\n      }\n    },\n    'completed.materials': function(){\n      if(this.completed.materials){\n        this.tab_index++;\n      }\n    }\n  }\n\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/new.vue?./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3413,7 +3425,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _com
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n\n /* harmony default export */ __webpack_exports__[\"default\"] = ({\n   data() {\n    return {\n      brands:{\n        supranet: 'supranet',\n        siemon: 'siemon'\n      },\n      translations:{\n        show: I18n.t('quotations.show'),\n        types: I18n.t('quotations.api_types')\n      },\n      quotation: {\n        client: {},\n        quotation_products: [],\n        quotation_services: [],\n        products_totals: {},\n        services_totals:{}\n      },\n      quotation_products: [],\n      quotation_products_headers: [],\n      quotation_services_headers: []\n    }\n  },\n  \n  mounted: function(){\n    this.getQuotation();\n    this.setQuotationServicesHeaders();\n  },\n\n  methods: {\n    \n    getQuotation: function(){\n      var id = this.$route.params.id;\n      this.http\n      .get(`/quotations/${id}`)\n      .then((response)=>{\n        if(response.successful){\n          this.quotation = response.data;\n          this.setQuotationProductsHeaders();\n        }else{\n          console.log(JSON.stringify(response));\n        }\n      }).catch((err)=>{\n        console.log(JSON.stringify(err));\n      });\n    },\n\n    setQuotationProductsHeaders: function(){\n      var headers = ['amount','material'];\n      var q_type = this.quotation.quotation_type;\n      //sets columns dependind on quotation_type, types are:\n      //t_comparative, t_supranet_only, t_siemon_only, t_simple\n      ['t_supranet_only','t_siemon_only'].forEach((type)=>{\n        if(q_type === 't_comparative' || q_type === type){\n          headers.push(`${type}_percent`);\n          //Adds four fields: unit price, total price, unit price plus percent and total price plus percent\n          ['','_with_percent'].forEach((suffix)=>{\n            headers = headers.concat([`${type}_price${suffix}`, `${type}_total${suffix}`]);\n          });\n        }\n      });\n      //simple type\n      if(q_type === 't_simple'){\n        headers = headers.concat(['brand','percent']);\n        ['','_with_percent'].forEach((suffix)=>{\n          headers = headers.concat([`${q_type}_price${suffix}`, `${q_type}_total${suffix}`]);\n        });\n      }\n      headers.forEach((element)=>{\n        this.quotation_products_headers.push({\n          key:element,\n          label:this.translations.show.headers[element],\n          sortable:true\n        });\n      });\n    },\n\n    setQuotationServicesHeaders: function(){\n      var headers = [\n        'amount',\n        'service',\n        'percent',\n        'price',\n        'total',\n        'price_with_percent',\n        'total_with_percent'\n      ].forEach((element)=>{\n        this.quotation_services_headers.push({\n          key:element,\n          label:this.translations.show.headers[element],\n          sortable:true\n        })\n      });\n    }\n  }\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/show.vue?./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _components_global_view_vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./components/global_view.vue */ \"./viewsrc/modules/quotations/components/global_view.vue\");\n\n\n\n /* harmony default export */ __webpack_exports__[\"default\"] = ({\n\n   components:{\n     'global-view': _components_global_view_vue__WEBPACK_IMPORTED_MODULE_0__[\"default\"]\n   }\n});\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/show.vue?./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3426,6 +3438,18 @@ eval("__webpack_require__.r(__webpack_exports__);\n\n /* harmony default export 
 
 "use strict";
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function() {\n  var _vm = this\n  var _h = _vm.$createElement\n  var _c = _vm._self._c || _h\n  return _c(\n    \"div\",\n    [\n      _c(\n        \"b-form\",\n        { on: { submit: _vm.updateConditions } },\n        [\n          _c(\"b-form-row\", [\n            _c(\n              \"div\",\n              { staticClass: \"col-12 mb-3\" },\n              [\n                _c(\n                  \"label\",\n                  { staticClass: \"mb-0 text-primary font-weight-bold\" },\n                  [\n                    _vm._v(\n                      \"\\n          \" +\n                        _vm._s(_vm.translations.titles.payment_condition) +\n                        \"\\n        \"\n                    )\n                  ]\n                ),\n                _vm._v(\" \"),\n                _c(\"b-textarea\", {\n                  model: {\n                    value: _vm.quotation.payment_condition,\n                    callback: function($$v) {\n                      _vm.$set(_vm.quotation, \"payment_condition\", $$v)\n                    },\n                    expression: \"quotation.payment_condition\"\n                  }\n                })\n              ],\n              1\n            ),\n            _vm._v(\" \"),\n            _c(\n              \"div\",\n              { staticClass: \"col-12 mb-3\" },\n              [\n                _c(\n                  \"label\",\n                  { staticClass: \"mb-0 text-primary font-weight-bold\" },\n                  [_vm._v(\" \" + _vm._s(_vm.translations.titles.credits) + \" \")]\n                ),\n                _vm._v(\" \"),\n                _c(\"b-textarea\", {\n                  model: {\n                    value: _vm.quotation.credits,\n                    callback: function($$v) {\n                      _vm.$set(_vm.quotation, \"credits\", $$v)\n                    },\n                    expression: \"quotation.credits\"\n                  }\n                })\n              ],\n              1\n            ),\n            _vm._v(\" \"),\n            _c(\n              \"div\",\n              { staticClass: \"col-12 mb-3\" },\n              [\n                _c(\n                  \"label\",\n                  { staticClass: \"mb-0 text-primary font-weight-bold\" },\n                  [_vm._v(\" \" + _vm._s(_vm.translations.titles.warranty) + \" \")]\n                ),\n                _vm._v(\" \"),\n                _c(\"b-textarea\", {\n                  model: {\n                    value: _vm.quotation.warranty,\n                    callback: function($$v) {\n                      _vm.$set(_vm.quotation, \"warranty\", $$v)\n                    },\n                    expression: \"quotation.warranty\"\n                  }\n                })\n              ],\n              1\n            ),\n            _vm._v(\" \"),\n            _c(\n              \"div\",\n              { staticClass: \"col-2 offset-10 mb-3\" },\n              [\n                _c(\n                  \"b-button\",\n                  { attrs: { variant: \"primary\", block: \"\", type: \"submit\" } },\n                  [\n                    _vm._v(\n                      \"\\n          \" +\n                        _vm._s(_vm.translations.next) +\n                        \"\\n        \"\n                    )\n                  ]\n                )\n              ],\n              1\n            )\n          ])\n        ],\n        1\n      )\n    ],\n    1\n  )\n}\nvar staticRenderFns = []\nrender._withStripped = true\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/conditions.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
+
+/***/ }),
+
+/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/components/global_view.vue?vue&type=template&id=96eec44c&":
+/*!****************************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./viewsrc/modules/quotations/components/global_view.vue?vue&type=template&id=96eec44c& ***!
+  \****************************************************************************************************************************************************************************************************************************/
+/*! exports provided: render, staticRenderFns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function() {\n  var _vm = this\n  var _h = _vm.$createElement\n  var _c = _vm._self._c || _h\n  return _c(\"div\", { staticClass: \"card\" }, [\n    _c(\"div\", { staticClass: \"card-header\" }, [\n      _c(\"div\", { staticClass: \"row\" }, [\n        _c(\"div\", { staticClass: \"col-6 mb-4\" }),\n        _vm._v(\" \"),\n        _c(\"div\", { staticClass: \"col-12 mb-4 text-right\" }, [\n          _c(\"h3\", [\n            _c(\"b\", [\n              _vm._v(\n                _vm._s(_vm.translations.show.titles.quotation_number) +\n                  \"\\n          \"\n              ),\n              _c(\"span\", { staticClass: \"text-danger\" }, [\n                _vm._v(\" \" + _vm._s(100 + _vm.quotation.id))\n              ])\n            ])\n          ])\n        ]),\n        _vm._v(\" \"),\n        _c(\"div\", { staticClass: \"col-7\" }, [\n          _c(\"h5\", [\n            _c(\"b\", [_vm._v(_vm._s(_vm.translations.show.titles.client.info))])\n          ]),\n          _vm._v(\" \"),\n          _c(\"div\", { staticClass: \"row\" }, [\n            _c(\"div\", { staticClass: \"col-12\" }, [\n              _c(\"b\", [\n                _vm._v(_vm._s(_vm.translations.show.titles.client.client))\n              ]),\n              _vm._v(\": \" + _vm._s(_vm.quotation.client.name) + \"\\n          \")\n            ]),\n            _vm._v(\" \"),\n            _c(\"div\", { staticClass: \"col-12\" }, [\n              _c(\"b\", [\n                _vm._v(_vm._s(_vm.translations.show.titles.client.address))\n              ]),\n              _vm._v(\n                \": \" + _vm._s(_vm.quotation.client.address) + \"\\n          \"\n              )\n            ]),\n            _vm._v(\" \"),\n            _c(\"div\", { staticClass: \"col-12\" }, [\n              _c(\"b\", [\n                _vm._v(_vm._s(_vm.translations.show.titles.client.nit))\n              ]),\n              _vm._v(\n                \": \" + _vm._s(_vm.quotation.client.nit) + \"\\n            \"\n              ),\n              _c(\"b\", [\n                _vm._v(_vm._s(_vm.translations.show.titles.client.phone))\n              ]),\n              _vm._v(\": \" + _vm._s(_vm.quotation.client.phone) + \"\\n          \")\n            ])\n          ])\n        ]),\n        _vm._v(\" \"),\n        _c(\"div\", { staticClass: \"col-5 text-right\" }, [\n          _c(\"h5\", [\n            _c(\"b\", [\n              _vm._v(_vm._s(_vm.translations.show.titles.quotation.info))\n            ])\n          ]),\n          _vm._v(\" \"),\n          _c(\"div\", { staticClass: \"row\" }, [\n            _c(\"div\", { staticClass: \"col-12\" }, [\n              _c(\"b\", [\n                _vm._v(_vm._s(_vm.translations.show.titles.quotation.number))\n              ]),\n              _vm._v(\": \" + _vm._s(100 + _vm.quotation.id) + \"\\n          \")\n            ]),\n            _vm._v(\" \"),\n            _c(\"div\", { staticClass: \"col-12\" }, [\n              _c(\"b\", [\n                _vm._v(_vm._s(_vm.translations.show.titles.quotation.date))\n              ]),\n              _vm._v(\n                \":\\n            \" +\n                  _vm._s(_vm.date.toLocalFormat(_vm.quotation.quotation_date)) +\n                  \"\\n          \"\n              )\n            ]),\n            _vm._v(\" \"),\n            _c(\"div\", { staticClass: \"col-12\" }, [\n              _c(\"b\", [\n                _vm._v(_vm._s(_vm.translations.show.titles.quotation.type))\n              ]),\n              _vm._v(\n                \":\\n            \" +\n                  _vm._s(_vm.translations.types[_vm.quotation.quotation_type]) +\n                  \"\\n          \"\n              )\n            ])\n          ])\n        ]),\n        _vm._v(\" \"),\n        _c(\n          \"div\",\n          { staticClass: \"col-12 mt-3\" },\n          [\n            _c(\"h5\", [\n              _c(\"b\", [_vm._v(_vm._s(_vm.translations.show.titles.materials))])\n            ]),\n            _vm._v(\" \"),\n            _c(\"b-table\", {\n              attrs: {\n                items: _vm.quotation.quotation_products,\n                fields: _vm.quotation_products_headers,\n                responsive: \"sm\",\n                \"thead-tr-class\": \"bg-primary text-white text-center\",\n                striped: \"\",\n                bordered: \"\",\n                small: \"\"\n              },\n              scopedSlots: _vm._u(\n                [\n                  _vm.quotation.quotation_type == \"t_comparative\"\n                    ? {\n                        key: \"thead-top\",\n                        fn: function(data) {\n                          return [\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"text-center\" },\n                              [\n                                _c(\"b-th\", {\n                                  staticClass: \"bg-primary text-white\",\n                                  attrs: { rowspan: \"3\", colspan: \"2\" }\n                                }),\n                                _vm._v(\" \"),\n                                _c(\n                                  \"b-th\",\n                                  {\n                                    staticClass: \"bg-dark\",\n                                    attrs: { colspan: \"5\" }\n                                  },\n                                  [\n                                    _vm._v(\n                                      _vm._s(\n                                        _vm.translations.show.custom_headers\n                                          .expression\n                                      )\n                                    )\n                                  ]\n                                ),\n                                _vm._v(\" \"),\n                                _c(\n                                  \"b-th\",\n                                  {\n                                    staticClass: \"bg-dark\",\n                                    attrs: { colspan: \"5\" }\n                                  },\n                                  [\n                                    _vm._v(\n                                      _vm._s(\n                                        _vm.translations.show.custom_headers\n                                          .expression\n                                      )\n                                    )\n                                  ]\n                                )\n                              ],\n                              1\n                            ),\n                            _vm._v(\" \"),\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"text-center text-danger\" },\n                              [\n                                _c(\"b-th\", { attrs: { colspan: \"5\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.brands\n                                        .t_supranet_only\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-th\", { attrs: { colspan: \"5\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.brands.t_siemon_only\n                                    )\n                                  )\n                                ])\n                              ],\n                              1\n                            ),\n                            _vm._v(\" \"),\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"text-center\" },\n                              [\n                                _c(\"b-th\"),\n                                _vm._v(\" \"),\n                                _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers\n                                        .without_percent\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers\n                                        .with_percent\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-th\"),\n                                _vm._v(\" \"),\n                                _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers\n                                        .without_percent\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers\n                                        .with_percent\n                                    )\n                                  )\n                                ])\n                              ],\n                              1\n                            )\n                          ]\n                        }\n                      }\n                    : {\n                        key: \"thead-top\",\n                        fn: function(data) {\n                          return [\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"text-center\" },\n                              [\n                                _vm.quotation.quotation_type == \"t_simple\"\n                                  ? _c(\"b-th\", {\n                                      staticClass: \"bg-primary text-white\",\n                                      attrs: { rowspan: \"2\", colspan: \"4\" }\n                                    })\n                                  : _c(\"b-th\", {\n                                      staticClass: \"bg-primary text-white\",\n                                      attrs: { rowspan: \"3\", colspan: \"3\" }\n                                    }),\n                                _vm._v(\" \"),\n                                _c(\n                                  \"b-th\",\n                                  {\n                                    staticClass: \"bg-dark\",\n                                    attrs: { colspan: \"4\" }\n                                  },\n                                  [\n                                    _vm._v(\n                                      _vm._s(\n                                        _vm.translations.show.custom_headers\n                                          .expression\n                                      )\n                                    )\n                                  ]\n                                )\n                              ],\n                              1\n                            ),\n                            _vm._v(\" \"),\n                            _vm.quotation.quotation_type !== \"t_simple\"\n                              ? _c(\n                                  \"b-tr\",\n                                  { staticClass: \"text-center text-danger\" },\n                                  [\n                                    _c(\"b-th\", { attrs: { colspan: \"4\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.brands[\n                                            _vm.quotation.quotation_type\n                                          ]\n                                        )\n                                      )\n                                    ])\n                                  ],\n                                  1\n                                )\n                              : _vm._e(),\n                            _vm._v(\" \"),\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"text-center\" },\n                              [\n                                _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers\n                                        .without_percent\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers\n                                        .with_percent\n                                    )\n                                  )\n                                ])\n                              ],\n                              1\n                            )\n                          ]\n                        }\n                      },\n                  _vm.quotation.quotation_type == \"t_comparative\"\n                    ? {\n                        key: \"custom-foot\",\n                        fn: function() {\n                          return [\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"bg-dark text-center\" },\n                              [\n                                _c(\"b-td\", { attrs: { colspan: \"2\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers.total\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    \"\\n                \" +\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          _vm.quotation.products_totals\n                                            .t_supranet_only.without_percent\n                                        )\n                                      ) +\n                                      \"\\n              \"\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    \"\\n                \" +\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          _vm.quotation.products_totals\n                                            .t_supranet_only.with_percent\n                                        )\n                                      ) +\n                                      \"\\n              \"\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    \"\\n                \" +\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          _vm.quotation.products_totals\n                                            .t_siemon_only.without_percent\n                                        )\n                                      ) +\n                                      \"\\n              \"\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    \"\\n                \" +\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          _vm.quotation.products_totals\n                                            .t_siemon_only.with_percent\n                                        )\n                                      ) +\n                                      \"\\n              \"\n                                  )\n                                ])\n                              ],\n                              1\n                            )\n                          ]\n                        },\n                        proxy: true\n                      }\n                    : {\n                        key: \"custom-foot\",\n                        fn: function() {\n                          return [\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"bg-dark text-center\" },\n                              [\n                                _vm.quotation.quotation_type == \"t_simple\"\n                                  ? _c(\"b-td\", { attrs: { colspan: \"4\" } }, [\n                                      _vm._v(\n                                        \"\\n                \" +\n                                          _vm._s(\n                                            _vm.translations.show.custom_headers\n                                              .total\n                                          ) +\n                                          \"\\n              \"\n                                      )\n                                    ])\n                                  : _c(\"b-td\", { attrs: { colspan: \"3\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .total\n                                        )\n                                      )\n                                    ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.currency.format(\n                                        _vm.quotation.products_totals\n                                          .without_percent\n                                      )\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.currency.format(\n                                        _vm.quotation.products_totals\n                                          .with_percent\n                                      )\n                                    )\n                                  )\n                                ])\n                              ],\n                              1\n                            )\n                          ]\n                        },\n                        proxy: true\n                      },\n                  {\n                    key: \"cell(amount)\",\n                    fn: function(data) {\n                      return [\n                        _c(\"div\", { staticClass: \"text-center\" }, [\n                          _vm._v(\n                            \"\\n              \" +\n                              _vm._s(parseInt(data.value)) +\n                              \"\\n            \"\n                          )\n                        ])\n                      ]\n                    }\n                  },\n                  {\n                    key: \"cell(brand)\",\n                    fn: function(data) {\n                      return [\n                        _c(\"div\", { staticClass: \"text-left\" }, [\n                          _vm._v(\n                            \"\\n              \" +\n                              _vm._s(data.value) +\n                              \"\\n            \"\n                          )\n                        ])\n                      ]\n                    }\n                  },\n                  {\n                    key: \"cell(material)\",\n                    fn: function(data) {\n                      return [\n                        _c(\"div\", { staticClass: \"text-left\" }, [\n                          _vm._v(\n                            \"\\n              \" +\n                              _vm._s(data.value) +\n                              \"\\n            \"\n                          )\n                        ])\n                      ]\n                    }\n                  },\n                  {\n                    key: \"cell()\",\n                    fn: function(data) {\n                      return [\n                        _c(\"div\", { staticClass: \"text-right\" }, [\n                          _vm._v(\n                            \"\\n              \" +\n                              _vm._s(_vm.currency.format(data.value)) +\n                              \"\\n            \"\n                          )\n                        ])\n                      ]\n                    }\n                  }\n                ],\n                null,\n                true\n              )\n            })\n          ],\n          1\n        ),\n        _vm._v(\" \"),\n        _c(\n          \"div\",\n          { staticClass: \"col-12 mt-2\" },\n          [\n            _c(\"h5\", [\n              _c(\"b\", [_vm._v(_vm._s(_vm.translations.show.titles.services))])\n            ]),\n            _vm._v(\" \"),\n            _vm.quotation.quotation_type == \"t_comparative\"\n              ? _c(\"h5\", [\n                  _c(\"b\", { staticClass: \"text-danger\" }, [\n                    _vm._v(_vm._s(_vm.translations.show.titles.option_a))\n                  ])\n                ])\n              : _vm._e(),\n            _vm._v(\" \"),\n            _c(\"b-table\", {\n              attrs: {\n                items: _vm.quotation.quotation_services,\n                fields: _vm.quotation_services_headers,\n                responsive: \"sm\",\n                \"thead-tr-class\": \"bg-primary text-white text-center\",\n                striped: \"\",\n                bordered: \"\",\n                small: \"\"\n              },\n              scopedSlots: _vm._u([\n                {\n                  key: \"thead-top\",\n                  fn: function(data) {\n                    return [\n                      _c(\n                        \"b-tr\",\n                        { staticClass: \"text-center\" },\n                        [\n                          _c(\"b-th\", {\n                            staticClass: \"bg-primary text-white\",\n                            attrs: { rowspan: \"3\", colspan: \"3\" }\n                          }),\n                          _vm._v(\" \"),\n                          _c(\n                            \"b-th\",\n                            { staticClass: \"bg-dark\", attrs: { colspan: \"4\" } },\n                            [\n                              _vm._v(\n                                _vm._s(\n                                  _vm.translations.show.custom_headers\n                                    .expression\n                                )\n                              )\n                            ]\n                          )\n                        ],\n                        1\n                      ),\n                      _vm._v(\" \"),\n                      _c(\n                        \"b-tr\",\n                        { staticClass: \"text-center text-danger\" },\n                        [\n                          _c(\"b-th\", { attrs: { colspan: \"4\" } }, [\n                            _vm._v(\n                              _vm._s(\n                                _vm.translations.show.custom_headers.prices\n                              )\n                            )\n                          ])\n                        ],\n                        1\n                      ),\n                      _vm._v(\" \"),\n                      _c(\n                        \"b-tr\",\n                        { staticClass: \"text-center\" },\n                        [\n                          _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                            _vm._v(\n                              _vm._s(\n                                _vm.translations.show.custom_headers\n                                  .without_percent\n                              )\n                            )\n                          ]),\n                          _vm._v(\" \"),\n                          _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                            _vm._v(\n                              _vm._s(\n                                _vm.translations.show.custom_headers\n                                  .with_percent\n                              )\n                            )\n                          ])\n                        ],\n                        1\n                      )\n                    ]\n                  }\n                },\n                {\n                  key: \"cell(amount)\",\n                  fn: function(data) {\n                    return [\n                      _c(\"div\", { staticClass: \"text-center\" }, [\n                        _vm._v(\n                          \"\\n              \" +\n                            _vm._s(parseInt(data.value)) +\n                            \"\\n            \"\n                        )\n                      ])\n                    ]\n                  }\n                },\n                {\n                  key: \"cell(service)\",\n                  fn: function(data) {\n                    return [\n                      _c(\"div\", { staticClass: \"text-left\" }, [\n                        _vm._v(\n                          \"\\n              \" +\n                            _vm._s(data.value) +\n                            \"\\n            \"\n                        )\n                      ])\n                    ]\n                  }\n                },\n                {\n                  key: \"cell()\",\n                  fn: function(data) {\n                    return [\n                      _c(\"div\", { staticClass: \"text-right\" }, [\n                        _vm._v(\n                          \"\\n              \" +\n                            _vm._s(_vm.currency.format(data.value)) +\n                            \"\\n            \"\n                        )\n                      ])\n                    ]\n                  }\n                },\n                {\n                  key: \"custom-foot\",\n                  fn: function() {\n                    return [\n                      _c(\n                        \"b-tr\",\n                        { staticClass: \"text-important\" },\n                        [\n                          _c(\"b-td\", { staticClass: \"text-center\" }, [\n                            _vm._v(\n                              _vm._s(\n                                _vm.translations.show.custom_footers.not_used\n                              )\n                            )\n                          ]),\n                          _vm._v(\" \"),\n                          _c(\"b-td\", [\n                            _vm._v(\n                              \"\\n                \" +\n                                _vm._s(\n                                  _vm.translations.show.custom_footers.materials\n                                ) +\n                                \"\\n                \"\n                            ),\n                            _vm.quotation.quotation_type == \"t_comparative\"\n                              ? _c(\"span\", [\n                                  _vm._v(\n                                    \"\\n                  \" +\n                                      _vm._s(\n                                        _vm.translations.show.brands\n                                          .t_supranet_only\n                                      ) +\n                                      \"\\n                \"\n                                  )\n                                ])\n                              : _c(\"span\", [\n                                  _vm._v(\n                                    \"\\n                  \" +\n                                      _vm._s(\n                                        _vm.translations.show.brands[\n                                          _vm.quotation.quotation_type\n                                        ]\n                                      ) +\n                                      \"\\n                \"\n                                  )\n                                ])\n                          ]),\n                          _vm._v(\" \"),\n                          _c(\"b-td\", { staticClass: \"text-center\" }, [\n                            _vm._v(\n                              _vm._s(\n                                _vm.translations.show.custom_footers.not_used\n                              )\n                            )\n                          ]),\n                          _vm._v(\" \"),\n                          _c(\"b-td\", { staticClass: \"text-center\" }, [\n                            _vm._v(\n                              _vm._s(\n                                _vm.translations.show.custom_footers.not_used\n                              )\n                            )\n                          ]),\n                          _vm._v(\" \"),\n                          _c(\"b-td\", { staticClass: \"text-right\" }, [\n                            _vm.quotation.quotation_type == \"t_comparative\"\n                              ? _c(\"span\", [\n                                  _vm._v(\n                                    \"\\n                  \" +\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          _vm.quotation.products_totals\n                                            .t_supranet_only.without_percent\n                                        )\n                                      ) +\n                                      \"\\n                \"\n                                  )\n                                ])\n                              : _c(\"span\", [\n                                  _vm._v(\n                                    \"\\n                  \" +\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          _vm.quotation.products_totals\n                                            .without_percent\n                                        )\n                                      ) +\n                                      \"\\n                \"\n                                  )\n                                ])\n                          ]),\n                          _vm._v(\" \"),\n                          _c(\"b-td\", { staticClass: \"text-center\" }, [\n                            _vm._v(\n                              _vm._s(\n                                _vm.translations.show.custom_footers.not_used\n                              )\n                            )\n                          ]),\n                          _vm._v(\" \"),\n                          _c(\"b-td\", { staticClass: \"text-right\" }, [\n                            _vm.quotation.quotation_type == \"t_comparative\"\n                              ? _c(\"span\", [\n                                  _vm._v(\n                                    \"\\n                  \" +\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          _vm.quotation.products_totals\n                                            .t_supranet_only.with_percent\n                                        )\n                                      ) +\n                                      \"\\n                \"\n                                  )\n                                ])\n                              : _c(\"span\", [\n                                  _vm._v(\n                                    \"\\n                  \" +\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          _vm.quotation.products_totals\n                                            .with_percent\n                                        )\n                                      ) +\n                                      \"\\n                \"\n                                  )\n                                ])\n                          ])\n                        ],\n                        1\n                      ),\n                      _vm._v(\" \"),\n                      _c(\n                        \"b-tr\",\n                        { staticClass: \"bg-dark text-center\" },\n                        [\n                          _c(\"b-td\", { attrs: { colspan: \"2\" } }, [\n                            _vm._v(\n                              _vm._s(_vm.translations.show.custom_headers.total)\n                            )\n                          ]),\n                          _vm._v(\" \"),\n                          _c(\"b-td\"),\n                          _vm._v(\" \"),\n                          _c(\"b-td\"),\n                          _vm._v(\" \"),\n                          _vm.quotation.quotation_type == \"t_comparative\"\n                            ? _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.currency.format(\n                                      parseFloat(\n                                        _vm.quotation.services_totals\n                                          .without_percent\n                                      ) +\n                                        parseFloat(\n                                          _vm.quotation.products_totals\n                                            .t_supranet_only.without_percent\n                                        )\n                                    )\n                                  ) + \"\\n              \"\n                                )\n                              ])\n                            : _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.currency.format(\n                                      parseFloat(\n                                        _vm.quotation.services_totals\n                                          .without_percent\n                                      ) +\n                                        parseFloat(\n                                          _vm.quotation.products_totals\n                                            .without_percent\n                                        )\n                                    )\n                                  ) + \"\\n              \"\n                                )\n                              ]),\n                          _vm._v(\" \"),\n                          _c(\"b-td\"),\n                          _vm._v(\" \"),\n                          _vm.quotation.quotation_type == \"t_comparative\"\n                            ? _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.currency.format(\n                                      parseFloat(\n                                        _vm.quotation.services_totals\n                                          .with_percent\n                                      ) +\n                                        parseFloat(\n                                          _vm.quotation.products_totals\n                                            .t_supranet_only.with_percent\n                                        )\n                                    )\n                                  ) + \"\\n              \"\n                                )\n                              ])\n                            : _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                _vm._v(\n                                  \"\\n                \" +\n                                    _vm._s(\n                                      _vm.currency.format(\n                                        parseFloat(\n                                          _vm.quotation.services_totals\n                                            .with_percent\n                                        ) +\n                                          parseFloat(\n                                            _vm.quotation.products_totals\n                                              .with_percent\n                                          )\n                                      )\n                                    ) +\n                                    \"\\n              \"\n                                )\n                              ])\n                        ],\n                        1\n                      )\n                    ]\n                  },\n                  proxy: true\n                }\n              ])\n            }),\n            _vm._v(\" \"),\n            _vm.quotation.quotation_type == \"t_comparative\"\n              ? _c(\"h5\", [\n                  _c(\"b\", { staticClass: \"text-danger\" }, [\n                    _vm._v(_vm._s(_vm.translations.show.titles.option_b))\n                  ])\n                ])\n              : _vm._e(),\n            _vm._v(\" \"),\n            _vm.quotation.quotation_type == \"t_comparative\"\n              ? _c(\"b-table\", {\n                  attrs: {\n                    items: _vm.quotation.quotation_services,\n                    fields: _vm.quotation_services_headers,\n                    responsive: \"sm\",\n                    \"thead-tr-class\": \"bg-primary text-white text-center\",\n                    striped: \"\",\n                    bordered: \"\",\n                    small: \"\"\n                  },\n                  scopedSlots: _vm._u(\n                    [\n                      {\n                        key: \"thead-top\",\n                        fn: function(data) {\n                          return [\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"text-center\" },\n                              [\n                                _c(\"b-th\", {\n                                  staticClass: \"bg-primary text-white\",\n                                  attrs: { rowspan: \"3\", colspan: \"3\" }\n                                }),\n                                _vm._v(\" \"),\n                                _c(\n                                  \"b-th\",\n                                  {\n                                    staticClass: \"bg-dark\",\n                                    attrs: { colspan: \"4\" }\n                                  },\n                                  [\n                                    _vm._v(\n                                      _vm._s(\n                                        _vm.translations.show.custom_headers\n                                          .expression\n                                      )\n                                    )\n                                  ]\n                                )\n                              ],\n                              1\n                            ),\n                            _vm._v(\" \"),\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"text-center text-danger\" },\n                              [\n                                _c(\"b-th\", { attrs: { colspan: \"4\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers\n                                        .prices\n                                    )\n                                  )\n                                ])\n                              ],\n                              1\n                            ),\n                            _vm._v(\" \"),\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"text-center\" },\n                              [\n                                _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers\n                                        .without_percent\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers\n                                        .with_percent\n                                    )\n                                  )\n                                ])\n                              ],\n                              1\n                            )\n                          ]\n                        }\n                      },\n                      {\n                        key: \"cell(amount)\",\n                        fn: function(data) {\n                          return [\n                            _c(\"div\", { staticClass: \"text-center\" }, [\n                              _vm._v(\n                                \"\\n              \" +\n                                  _vm._s(parseInt(data.value)) +\n                                  \"\\n            \"\n                              )\n                            ])\n                          ]\n                        }\n                      },\n                      {\n                        key: \"cell(service)\",\n                        fn: function(data) {\n                          return [\n                            _c(\"div\", { staticClass: \"text-left\" }, [\n                              _vm._v(\n                                \"\\n              \" +\n                                  _vm._s(data.value) +\n                                  \"\\n            \"\n                              )\n                            ])\n                          ]\n                        }\n                      },\n                      {\n                        key: \"cell()\",\n                        fn: function(data) {\n                          return [\n                            _c(\"div\", { staticClass: \"text-right\" }, [\n                              _vm._v(\n                                \"\\n              \" +\n                                  _vm._s(_vm.currency.format(data.value)) +\n                                  \"\\n            \"\n                              )\n                            ])\n                          ]\n                        }\n                      },\n                      {\n                        key: \"custom-foot\",\n                        fn: function() {\n                          return [\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"text-important\" },\n                              [\n                                _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_footers\n                                        .not_used\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", [\n                                  _vm._v(\n                                    \"\\n                \" +\n                                      _vm._s(\n                                        _vm.translations.show.custom_footers\n                                          .materials\n                                      ) +\n                                      \"\\n                \"\n                                  ),\n                                  _vm.quotation.quotation_type ==\n                                  \"t_comparative\"\n                                    ? _c(\"span\", [\n                                        _vm._v(\n                                          \"\\n                  \" +\n                                            _vm._s(\n                                              _vm.translations.show.brands\n                                                .t_siemon_only\n                                            ) +\n                                            \"\\n                \"\n                                        )\n                                      ])\n                                    : _c(\"span\", [\n                                        _vm._v(\n                                          \"\\n                  \" +\n                                            _vm._s(\n                                              _vm.translations.show.brands[\n                                                _vm.quotation.quotation_type\n                                              ]\n                                            ) +\n                                            \"\\n                \"\n                                        )\n                                      ])\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_footers\n                                        .not_used\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_footers\n                                        .not_used\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    \"\\n                \" +\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          _vm.quotation.products_totals\n                                            .t_siemon_only.without_percent\n                                        )\n                                      ) +\n                                      \"\\n              \"\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_footers\n                                        .not_used\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    \"\\n                \" +\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          _vm.quotation.products_totals\n                                            .t_siemon_only.with_percent\n                                        )\n                                      ) +\n                                      \"\\n              \"\n                                  )\n                                ])\n                              ],\n                              1\n                            ),\n                            _vm._v(\" \"),\n                            _c(\n                              \"b-tr\",\n                              { staticClass: \"bg-dark text-center\" },\n                              [\n                                _c(\"b-td\", { attrs: { colspan: \"2\" } }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers.total\n                                    )\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.currency.format(\n                                        parseFloat(\n                                          _vm.quotation.services_totals\n                                            .without_percent\n                                        ) +\n                                          parseFloat(\n                                            _vm.quotation.products_totals\n                                              .t_siemon_only.without_percent\n                                          )\n                                      )\n                                    ) + \"\\n              \"\n                                  )\n                                ]),\n                                _vm._v(\" \"),\n                                _c(\"b-td\"),\n                                _vm._v(\" \"),\n                                _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.currency.format(\n                                        parseFloat(\n                                          _vm.quotation.services_totals\n                                            .with_percent\n                                        ) +\n                                          parseFloat(\n                                            _vm.quotation.products_totals\n                                              .t_siemon_only.with_percent\n                                          )\n                                      )\n                                    ) + \"\\n              \"\n                                  )\n                                ])\n                              ],\n                              1\n                            )\n                          ]\n                        },\n                        proxy: true\n                      }\n                    ],\n                    null,\n                    false,\n                    85357279\n                  )\n                })\n              : _vm._e()\n          ],\n          1\n        ),\n        _vm._v(\" \"),\n        _c(\"div\", { staticClass: \"col-12 mt-2\" }, [\n          _c(\"h5\", [\n            _c(\"b\", [_vm._v(_vm._s(_vm.translations.show.titles.credits))])\n          ]),\n          _vm._v(\" \"),\n          _c(\"span\", { staticClass: \"text-justify\" }, [\n            _vm._v(\n              \"\\n          \" + _vm._s(_vm.quotation.credits) + \"\\n        \"\n            )\n          ])\n        ]),\n        _vm._v(\" \"),\n        _c(\"div\", { staticClass: \"col-12 mt-2\" }, [\n          _c(\"h5\", [\n            _c(\"b\", [_vm._v(_vm._s(_vm.translations.show.titles.warranty))])\n          ]),\n          _vm._v(\" \"),\n          _c(\"span\", { staticClass: \"text-justify\" }, [\n            _vm._v(\n              \"\\n          \" + _vm._s(_vm.quotation.warranty) + \"\\n        \"\n            )\n          ])\n        ]),\n        _vm._v(\" \"),\n        _c(\"div\", { staticClass: \"col-12 mt-2\" }, [\n          _c(\"h5\", [\n            _c(\"b\", [\n              _vm._v(_vm._s(_vm.translations.show.titles.payment_condition))\n            ])\n          ]),\n          _vm._v(\" \"),\n          _c(\"span\", { staticClass: \"text-justify\" }, [\n            _vm._v(\n              \"\\n          \" +\n                _vm._s(_vm.quotation.payment_condition) +\n                \"\\n        \"\n            )\n          ])\n        ])\n      ])\n    ])\n  ])\n}\nvar staticRenderFns = []\nrender._withStripped = true\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/global_view.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3497,7 +3521,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function() {\n  var _vm = this\n  var _h = _vm.$createElement\n  var _c = _vm._self._c || _h\n  return _c(\"div\", { staticClass: \"row\" }, [\n    _c(\"div\", { staticClass: \"col-lg-12 offset-xl-1 col-xl-10\" }, [\n      _c(\"div\", { staticClass: \"card\" }, [\n        _c(\"div\", { staticClass: \"card-header\" }, [\n          _c(\"div\", { staticClass: \"row\" }, [\n            _c(\"div\", { staticClass: \"col-6 mb-4\" }),\n            _vm._v(\" \"),\n            _c(\"div\", { staticClass: \"col-12 mb-4 text-right\" }, [\n              _c(\"h3\", [\n                _c(\"b\", [\n                  _vm._v(\n                    _vm._s(_vm.translations.show.titles.quotation_number) +\n                      \"\\n              \"\n                  ),\n                  _c(\"span\", { staticClass: \"text-danger\" }, [\n                    _vm._v(\" \" + _vm._s(100 + _vm.quotation.id))\n                  ])\n                ])\n              ])\n            ]),\n            _vm._v(\" \"),\n            _c(\"div\", { staticClass: \"col-7\" }, [\n              _c(\"h5\", [\n                _c(\"b\", [\n                  _vm._v(_vm._s(_vm.translations.show.titles.client.info))\n                ])\n              ]),\n              _vm._v(\" \"),\n              _c(\"div\", { staticClass: \"row\" }, [\n                _c(\"div\", { staticClass: \"col-12\" }, [\n                  _c(\"b\", [\n                    _vm._v(_vm._s(_vm.translations.show.titles.client.client))\n                  ]),\n                  _vm._v(\n                    \": \" +\n                      _vm._s(_vm.quotation.client.name) +\n                      \"\\n              \"\n                  )\n                ]),\n                _vm._v(\" \"),\n                _c(\"div\", { staticClass: \"col-12\" }, [\n                  _c(\"b\", [\n                    _vm._v(_vm._s(_vm.translations.show.titles.client.address))\n                  ]),\n                  _vm._v(\n                    \": \" +\n                      _vm._s(_vm.quotation.client.address) +\n                      \"\\n              \"\n                  )\n                ]),\n                _vm._v(\" \"),\n                _c(\"div\", { staticClass: \"col-12\" }, [\n                  _c(\"b\", [\n                    _vm._v(_vm._s(_vm.translations.show.titles.client.nit))\n                  ]),\n                  _vm._v(\n                    \": \" +\n                      _vm._s(_vm.quotation.client.nit) +\n                      \"\\n                \"\n                  ),\n                  _c(\"b\", [\n                    _vm._v(_vm._s(_vm.translations.show.titles.client.phone))\n                  ]),\n                  _vm._v(\n                    \": \" +\n                      _vm._s(_vm.quotation.client.phone) +\n                      \"\\n              \"\n                  )\n                ])\n              ])\n            ]),\n            _vm._v(\" \"),\n            _c(\"div\", { staticClass: \"col-5 text-right\" }, [\n              _c(\"h5\", [\n                _c(\"b\", [\n                  _vm._v(_vm._s(_vm.translations.show.titles.quotation.info))\n                ])\n              ]),\n              _vm._v(\" \"),\n              _c(\"div\", { staticClass: \"row\" }, [\n                _c(\"div\", { staticClass: \"col-12\" }, [\n                  _c(\"b\", [\n                    _vm._v(\n                      _vm._s(_vm.translations.show.titles.quotation.number)\n                    )\n                  ]),\n                  _vm._v(\n                    \": \" + _vm._s(100 + _vm.quotation.id) + \"\\n              \"\n                  )\n                ]),\n                _vm._v(\" \"),\n                _c(\"div\", { staticClass: \"col-12\" }, [\n                  _c(\"b\", [\n                    _vm._v(_vm._s(_vm.translations.show.titles.quotation.date))\n                  ]),\n                  _vm._v(\n                    \":\\n                \" +\n                      _vm._s(\n                        _vm.date.toLocalFormat(_vm.quotation.quotation_date)\n                      ) +\n                      \"\\n              \"\n                  )\n                ]),\n                _vm._v(\" \"),\n                _c(\"div\", { staticClass: \"col-12\" }, [\n                  _c(\"b\", [\n                    _vm._v(_vm._s(_vm.translations.show.titles.quotation.type))\n                  ]),\n                  _vm._v(\n                    \":\\n                \" +\n                      _vm._s(\n                        _vm.translations.types[_vm.quotation.quotation_type]\n                      ) +\n                      \"\\n              \"\n                  )\n                ])\n              ])\n            ]),\n            _vm._v(\" \"),\n            _c(\n              \"div\",\n              { staticClass: \"col-12 mt-3\" },\n              [\n                _c(\"h5\", [\n                  _c(\"b\", [\n                    _vm._v(_vm._s(_vm.translations.show.titles.materials))\n                  ])\n                ]),\n                _vm._v(\" \"),\n                _c(\"b-table\", {\n                  attrs: {\n                    items: _vm.quotation.quotation_products,\n                    fields: _vm.quotation_products_headers,\n                    responsive: \"sm\",\n                    \"thead-tr-class\": \"bg-primary text-white text-center\",\n                    striped: \"\",\n                    bordered: \"\",\n                    small: \"\"\n                  },\n                  scopedSlots: _vm._u(\n                    [\n                      _vm.quotation.quotation_type == \"t_comparative\"\n                        ? {\n                            key: \"thead-top\",\n                            fn: function(data) {\n                              return [\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"text-center\" },\n                                  [\n                                    _c(\"b-th\", {\n                                      staticClass: \"bg-primary text-white\",\n                                      attrs: { rowspan: \"3\", colspan: \"2\" }\n                                    }),\n                                    _vm._v(\" \"),\n                                    _c(\n                                      \"b-th\",\n                                      {\n                                        staticClass: \"bg-dark\",\n                                        attrs: { colspan: \"5\" }\n                                      },\n                                      [\n                                        _vm._v(\n                                          _vm._s(\n                                            _vm.translations.show.custom_headers\n                                              .expression\n                                          )\n                                        )\n                                      ]\n                                    ),\n                                    _vm._v(\" \"),\n                                    _c(\n                                      \"b-th\",\n                                      {\n                                        staticClass: \"bg-dark\",\n                                        attrs: { colspan: \"5\" }\n                                      },\n                                      [\n                                        _vm._v(\n                                          _vm._s(\n                                            _vm.translations.show.custom_headers\n                                              .expression\n                                          )\n                                        )\n                                      ]\n                                    )\n                                  ],\n                                  1\n                                ),\n                                _vm._v(\" \"),\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"text-center text-danger\" },\n                                  [\n                                    _c(\"b-th\", { attrs: { colspan: \"5\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.brands\n                                            .t_supranet_only\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-th\", { attrs: { colspan: \"5\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.brands\n                                            .t_siemon_only\n                                        )\n                                      )\n                                    ])\n                                  ],\n                                  1\n                                ),\n                                _vm._v(\" \"),\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"text-center\" },\n                                  [\n                                    _c(\"b-th\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .without_percent\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .with_percent\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-th\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .without_percent\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .with_percent\n                                        )\n                                      )\n                                    ])\n                                  ],\n                                  1\n                                )\n                              ]\n                            }\n                          }\n                        : {\n                            key: \"thead-top\",\n                            fn: function(data) {\n                              return [\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"text-center\" },\n                                  [\n                                    _vm.quotation.quotation_type == \"t_simple\"\n                                      ? _c(\"b-th\", {\n                                          staticClass: \"bg-primary text-white\",\n                                          attrs: { rowspan: \"2\", colspan: \"4\" }\n                                        })\n                                      : _c(\"b-th\", {\n                                          staticClass: \"bg-primary text-white\",\n                                          attrs: { rowspan: \"3\", colspan: \"3\" }\n                                        }),\n                                    _vm._v(\" \"),\n                                    _c(\n                                      \"b-th\",\n                                      {\n                                        staticClass: \"bg-dark\",\n                                        attrs: { colspan: \"4\" }\n                                      },\n                                      [\n                                        _vm._v(\n                                          _vm._s(\n                                            _vm.translations.show.custom_headers\n                                              .expression\n                                          )\n                                        )\n                                      ]\n                                    )\n                                  ],\n                                  1\n                                ),\n                                _vm._v(\" \"),\n                                _vm.quotation.quotation_type !== \"t_simple\"\n                                  ? _c(\n                                      \"b-tr\",\n                                      {\n                                        staticClass: \"text-center text-danger\"\n                                      },\n                                      [\n                                        _c(\n                                          \"b-th\",\n                                          { attrs: { colspan: \"4\" } },\n                                          [\n                                            _vm._v(\n                                              _vm._s(\n                                                _vm.translations.show.brands[\n                                                  _vm.quotation.quotation_type\n                                                ]\n                                              )\n                                            )\n                                          ]\n                                        )\n                                      ],\n                                      1\n                                    )\n                                  : _vm._e(),\n                                _vm._v(\" \"),\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"text-center\" },\n                                  [\n                                    _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .without_percent\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .with_percent\n                                        )\n                                      )\n                                    ])\n                                  ],\n                                  1\n                                )\n                              ]\n                            }\n                          },\n                      _vm.quotation.quotation_type == \"t_comparative\"\n                        ? {\n                            key: \"custom-foot\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"bg-dark text-center\" },\n                                  [\n                                    _c(\"b-td\", { attrs: { colspan: \"2\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .total\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                      _vm._v(\n                                        \"\\n                    \" +\n                                          _vm._s(\n                                            _vm.currency.format(\n                                              _vm.quotation.products_totals\n                                                .t_supranet_only.without_percent\n                                            )\n                                          ) +\n                                          \"\\n                  \"\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                      _vm._v(\n                                        \"\\n                    \" +\n                                          _vm._s(\n                                            _vm.currency.format(\n                                              _vm.quotation.products_totals\n                                                .t_supranet_only.with_percent\n                                            )\n                                          ) +\n                                          \"\\n                  \"\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                      _vm._v(\n                                        \"\\n                    \" +\n                                          _vm._s(\n                                            _vm.currency.format(\n                                              _vm.quotation.products_totals\n                                                .t_siemon_only.without_percent\n                                            )\n                                          ) +\n                                          \"\\n                  \"\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                      _vm._v(\n                                        \"\\n                    \" +\n                                          _vm._s(\n                                            _vm.currency.format(\n                                              _vm.quotation.products_totals\n                                                .t_siemon_only.with_percent\n                                            )\n                                          ) +\n                                          \"\\n                  \"\n                                      )\n                                    ])\n                                  ],\n                                  1\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        : {\n                            key: \"custom-foot\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"bg-dark text-center\" },\n                                  [\n                                    _vm.quotation.quotation_type == \"t_simple\"\n                                      ? _c(\n                                          \"b-td\",\n                                          { attrs: { colspan: \"4\" } },\n                                          [\n                                            _vm._v(\n                                              \"\\n                    \" +\n                                                _vm._s(\n                                                  _vm.translations.show\n                                                    .custom_headers.total\n                                                ) +\n                                                \"\\n                  \"\n                                            )\n                                          ]\n                                        )\n                                      : _c(\n                                          \"b-td\",\n                                          { attrs: { colspan: \"3\" } },\n                                          [\n                                            _vm._v(\n                                              _vm._s(\n                                                _vm.translations.show\n                                                  .custom_headers.total\n                                              )\n                                            )\n                                          ]\n                                        ),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.currency.format(\n                                            _vm.quotation.products_totals\n                                              .without_percent\n                                          )\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.currency.format(\n                                            _vm.quotation.products_totals\n                                              .with_percent\n                                          )\n                                        )\n                                      )\n                                    ])\n                                  ],\n                                  1\n                                )\n                              ]\n                            },\n                            proxy: true\n                          },\n                      {\n                        key: \"cell(amount)\",\n                        fn: function(data) {\n                          return [\n                            _c(\"div\", { staticClass: \"text-center\" }, [\n                              _vm._v(\n                                \"\\n                  \" +\n                                  _vm._s(parseInt(data.value)) +\n                                  \"\\n                \"\n                              )\n                            ])\n                          ]\n                        }\n                      },\n                      {\n                        key: \"cell(brand)\",\n                        fn: function(data) {\n                          return [\n                            _c(\"div\", { staticClass: \"text-left\" }, [\n                              _vm._v(\n                                \"\\n                  \" +\n                                  _vm._s(data.value) +\n                                  \"\\n                \"\n                              )\n                            ])\n                          ]\n                        }\n                      },\n                      {\n                        key: \"cell(material)\",\n                        fn: function(data) {\n                          return [\n                            _c(\"div\", { staticClass: \"text-left\" }, [\n                              _vm._v(\n                                \"\\n                  \" +\n                                  _vm._s(data.value) +\n                                  \"\\n                \"\n                              )\n                            ])\n                          ]\n                        }\n                      },\n                      {\n                        key: \"cell()\",\n                        fn: function(data) {\n                          return [\n                            _c(\"div\", { staticClass: \"text-right\" }, [\n                              _vm._v(\n                                \"\\n                  \" +\n                                  _vm._s(_vm.currency.format(data.value)) +\n                                  \"\\n                \"\n                              )\n                            ])\n                          ]\n                        }\n                      }\n                    ],\n                    null,\n                    true\n                  )\n                })\n              ],\n              1\n            ),\n            _vm._v(\" \"),\n            _c(\n              \"div\",\n              { staticClass: \"col-12 mt-2\" },\n              [\n                _c(\"h5\", [\n                  _c(\"b\", [\n                    _vm._v(_vm._s(_vm.translations.show.titles.services))\n                  ])\n                ]),\n                _vm._v(\" \"),\n                _vm.quotation.quotation_type == \"t_comparative\"\n                  ? _c(\"h5\", [\n                      _c(\"b\", { staticClass: \"text-danger\" }, [\n                        _vm._v(_vm._s(_vm.translations.show.titles.option_a))\n                      ])\n                    ])\n                  : _vm._e(),\n                _vm._v(\" \"),\n                _c(\"b-table\", {\n                  attrs: {\n                    items: _vm.quotation.quotation_services,\n                    fields: _vm.quotation_services_headers,\n                    responsive: \"sm\",\n                    \"thead-tr-class\": \"bg-primary text-white text-center\",\n                    striped: \"\",\n                    bordered: \"\",\n                    small: \"\"\n                  },\n                  scopedSlots: _vm._u([\n                    {\n                      key: \"thead-top\",\n                      fn: function(data) {\n                        return [\n                          _c(\n                            \"b-tr\",\n                            { staticClass: \"text-center\" },\n                            [\n                              _c(\"b-th\", {\n                                staticClass: \"bg-primary text-white\",\n                                attrs: { rowspan: \"3\", colspan: \"3\" }\n                              }),\n                              _vm._v(\" \"),\n                              _c(\n                                \"b-th\",\n                                {\n                                  staticClass: \"bg-dark\",\n                                  attrs: { colspan: \"4\" }\n                                },\n                                [\n                                  _vm._v(\n                                    _vm._s(\n                                      _vm.translations.show.custom_headers\n                                        .expression\n                                    )\n                                  )\n                                ]\n                              )\n                            ],\n                            1\n                          ),\n                          _vm._v(\" \"),\n                          _c(\n                            \"b-tr\",\n                            { staticClass: \"text-center text-danger\" },\n                            [\n                              _c(\"b-th\", { attrs: { colspan: \"4\" } }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.translations.show.custom_headers.prices\n                                  )\n                                )\n                              ])\n                            ],\n                            1\n                          ),\n                          _vm._v(\" \"),\n                          _c(\n                            \"b-tr\",\n                            { staticClass: \"text-center\" },\n                            [\n                              _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.translations.show.custom_headers\n                                      .without_percent\n                                  )\n                                )\n                              ]),\n                              _vm._v(\" \"),\n                              _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.translations.show.custom_headers\n                                      .with_percent\n                                  )\n                                )\n                              ])\n                            ],\n                            1\n                          )\n                        ]\n                      }\n                    },\n                    {\n                      key: \"cell(amount)\",\n                      fn: function(data) {\n                        return [\n                          _c(\"div\", { staticClass: \"text-center\" }, [\n                            _vm._v(\n                              \"\\n                  \" +\n                                _vm._s(parseInt(data.value)) +\n                                \"\\n                \"\n                            )\n                          ])\n                        ]\n                      }\n                    },\n                    {\n                      key: \"cell(service)\",\n                      fn: function(data) {\n                        return [\n                          _c(\"div\", { staticClass: \"text-left\" }, [\n                            _vm._v(\n                              \"\\n                  \" +\n                                _vm._s(data.value) +\n                                \"\\n                \"\n                            )\n                          ])\n                        ]\n                      }\n                    },\n                    {\n                      key: \"cell()\",\n                      fn: function(data) {\n                        return [\n                          _c(\"div\", { staticClass: \"text-right\" }, [\n                            _vm._v(\n                              \"\\n                  \" +\n                                _vm._s(_vm.currency.format(data.value)) +\n                                \"\\n                \"\n                            )\n                          ])\n                        ]\n                      }\n                    },\n                    {\n                      key: \"custom-foot\",\n                      fn: function() {\n                        return [\n                          _c(\n                            \"b-tr\",\n                            { staticClass: \"text-important\" },\n                            [\n                              _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.translations.show.custom_footers\n                                      .not_used\n                                  )\n                                )\n                              ]),\n                              _vm._v(\" \"),\n                              _c(\"b-td\", [\n                                _vm._v(\n                                  \"\\n                    \" +\n                                    _vm._s(\n                                      _vm.translations.show.custom_footers\n                                        .materials\n                                    ) +\n                                    \"\\n                    \"\n                                ),\n                                _vm.quotation.quotation_type == \"t_comparative\"\n                                  ? _c(\"span\", [\n                                      _vm._v(\n                                        \"\\n                      \" +\n                                          _vm._s(\n                                            _vm.translations.show.brands\n                                              .t_supranet_only\n                                          ) +\n                                          \"\\n                    \"\n                                      )\n                                    ])\n                                  : _c(\"span\", [\n                                      _vm._v(\n                                        \"\\n                      \" +\n                                          _vm._s(\n                                            _vm.translations.show.brands[\n                                              _vm.quotation.quotation_type\n                                            ]\n                                          ) +\n                                          \"\\n                    \"\n                                      )\n                                    ])\n                              ]),\n                              _vm._v(\" \"),\n                              _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.translations.show.custom_footers\n                                      .not_used\n                                  )\n                                )\n                              ]),\n                              _vm._v(\" \"),\n                              _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.translations.show.custom_footers\n                                      .not_used\n                                  )\n                                )\n                              ]),\n                              _vm._v(\" \"),\n                              _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                _vm.quotation.quotation_type == \"t_comparative\"\n                                  ? _c(\"span\", [\n                                      _vm._v(\n                                        \"\\n                      \" +\n                                          _vm._s(\n                                            _vm.currency.format(\n                                              _vm.quotation.products_totals\n                                                .t_supranet_only.without_percent\n                                            )\n                                          ) +\n                                          \"\\n                    \"\n                                      )\n                                    ])\n                                  : _c(\"span\", [\n                                      _vm._v(\n                                        \"\\n                      \" +\n                                          _vm._s(\n                                            _vm.currency.format(\n                                              _vm.quotation.products_totals\n                                                .without_percent\n                                            )\n                                          ) +\n                                          \"\\n                    \"\n                                      )\n                                    ])\n                              ]),\n                              _vm._v(\" \"),\n                              _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.translations.show.custom_footers\n                                      .not_used\n                                  )\n                                )\n                              ]),\n                              _vm._v(\" \"),\n                              _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                _vm.quotation.quotation_type == \"t_comparative\"\n                                  ? _c(\"span\", [\n                                      _vm._v(\n                                        \"\\n                      \" +\n                                          _vm._s(\n                                            _vm.currency.format(\n                                              _vm.quotation.products_totals\n                                                .t_supranet_only.with_percent\n                                            )\n                                          ) +\n                                          \"\\n                    \"\n                                      )\n                                    ])\n                                  : _c(\"span\", [\n                                      _vm._v(\n                                        \"\\n                      \" +\n                                          _vm._s(\n                                            _vm.currency.format(\n                                              _vm.quotation.products_totals\n                                                .with_percent\n                                            )\n                                          ) +\n                                          \"\\n                    \"\n                                      )\n                                    ])\n                              ])\n                            ],\n                            1\n                          ),\n                          _vm._v(\" \"),\n                          _c(\n                            \"b-tr\",\n                            { staticClass: \"bg-dark text-center\" },\n                            [\n                              _c(\"b-td\", { attrs: { colspan: \"2\" } }, [\n                                _vm._v(\n                                  _vm._s(\n                                    _vm.translations.show.custom_headers.total\n                                  )\n                                )\n                              ]),\n                              _vm._v(\" \"),\n                              _c(\"b-td\"),\n                              _vm._v(\" \"),\n                              _c(\"b-td\"),\n                              _vm._v(\" \"),\n                              _vm.quotation.quotation_type == \"t_comparative\"\n                                ? _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                    _vm._v(\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          parseFloat(\n                                            _vm.quotation.services_totals\n                                              .without_percent\n                                          ) +\n                                            parseFloat(\n                                              _vm.quotation.products_totals\n                                                .t_supranet_only.without_percent\n                                            )\n                                        )\n                                      ) + \"\\n                  \"\n                                    )\n                                  ])\n                                : _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                    _vm._v(\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          parseFloat(\n                                            _vm.quotation.services_totals\n                                              .without_percent\n                                          ) +\n                                            parseFloat(\n                                              _vm.quotation.products_totals\n                                                .without_percent\n                                            )\n                                        )\n                                      ) + \"\\n                  \"\n                                    )\n                                  ]),\n                              _vm._v(\" \"),\n                              _c(\"b-td\"),\n                              _vm._v(\" \"),\n                              _vm.quotation.quotation_type == \"t_comparative\"\n                                ? _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                    _vm._v(\n                                      _vm._s(\n                                        _vm.currency.format(\n                                          parseFloat(\n                                            _vm.quotation.services_totals\n                                              .with_percent\n                                          ) +\n                                            parseFloat(\n                                              _vm.quotation.products_totals\n                                                .t_supranet_only.with_percent\n                                            )\n                                        )\n                                      ) + \"\\n                  \"\n                                    )\n                                  ])\n                                : _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                    _vm._v(\n                                      \"\\n                    \" +\n                                        _vm._s(\n                                          _vm.currency.format(\n                                            parseFloat(\n                                              _vm.quotation.services_totals\n                                                .with_percent\n                                            ) +\n                                              parseFloat(\n                                                _vm.quotation.products_totals\n                                                  .with_percent\n                                              )\n                                          )\n                                        ) +\n                                        \"\\n                  \"\n                                    )\n                                  ])\n                            ],\n                            1\n                          )\n                        ]\n                      },\n                      proxy: true\n                    }\n                  ])\n                }),\n                _vm._v(\" \"),\n                _vm.quotation.quotation_type == \"t_comparative\"\n                  ? _c(\"h5\", [\n                      _c(\"b\", { staticClass: \"text-danger\" }, [\n                        _vm._v(_vm._s(_vm.translations.show.titles.option_b))\n                      ])\n                    ])\n                  : _vm._e(),\n                _vm._v(\" \"),\n                _vm.quotation.quotation_type == \"t_comparative\"\n                  ? _c(\"b-table\", {\n                      attrs: {\n                        items: _vm.quotation.quotation_services,\n                        fields: _vm.quotation_services_headers,\n                        responsive: \"sm\",\n                        \"thead-tr-class\": \"bg-primary text-white text-center\",\n                        striped: \"\",\n                        bordered: \"\",\n                        small: \"\"\n                      },\n                      scopedSlots: _vm._u(\n                        [\n                          {\n                            key: \"thead-top\",\n                            fn: function(data) {\n                              return [\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"text-center\" },\n                                  [\n                                    _c(\"b-th\", {\n                                      staticClass: \"bg-primary text-white\",\n                                      attrs: { rowspan: \"3\", colspan: \"3\" }\n                                    }),\n                                    _vm._v(\" \"),\n                                    _c(\n                                      \"b-th\",\n                                      {\n                                        staticClass: \"bg-dark\",\n                                        attrs: { colspan: \"4\" }\n                                      },\n                                      [\n                                        _vm._v(\n                                          _vm._s(\n                                            _vm.translations.show.custom_headers\n                                              .expression\n                                          )\n                                        )\n                                      ]\n                                    )\n                                  ],\n                                  1\n                                ),\n                                _vm._v(\" \"),\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"text-center text-danger\" },\n                                  [\n                                    _c(\"b-th\", { attrs: { colspan: \"4\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .prices\n                                        )\n                                      )\n                                    ])\n                                  ],\n                                  1\n                                ),\n                                _vm._v(\" \"),\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"text-center\" },\n                                  [\n                                    _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .without_percent\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-th\", { attrs: { colspan: \"2\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .with_percent\n                                        )\n                                      )\n                                    ])\n                                  ],\n                                  1\n                                )\n                              ]\n                            }\n                          },\n                          {\n                            key: \"cell(amount)\",\n                            fn: function(data) {\n                              return [\n                                _c(\"div\", { staticClass: \"text-center\" }, [\n                                  _vm._v(\n                                    \"\\n                  \" +\n                                      _vm._s(parseInt(data.value)) +\n                                      \"\\n                \"\n                                  )\n                                ])\n                              ]\n                            }\n                          },\n                          {\n                            key: \"cell(service)\",\n                            fn: function(data) {\n                              return [\n                                _c(\"div\", { staticClass: \"text-left\" }, [\n                                  _vm._v(\n                                    \"\\n                  \" +\n                                      _vm._s(data.value) +\n                                      \"\\n                \"\n                                  )\n                                ])\n                              ]\n                            }\n                          },\n                          {\n                            key: \"cell()\",\n                            fn: function(data) {\n                              return [\n                                _c(\"div\", { staticClass: \"text-right\" }, [\n                                  _vm._v(\n                                    \"\\n                  \" +\n                                      _vm._s(_vm.currency.format(data.value)) +\n                                      \"\\n                \"\n                                  )\n                                ])\n                              ]\n                            }\n                          },\n                          {\n                            key: \"custom-foot\",\n                            fn: function() {\n                              return [\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"text-important\" },\n                                  [\n                                    _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_footers\n                                            .not_used\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", [\n                                      _vm._v(\n                                        \"\\n                    \" +\n                                          _vm._s(\n                                            _vm.translations.show.custom_footers\n                                              .materials\n                                          ) +\n                                          \"\\n                    \"\n                                      ),\n                                      _vm.quotation.quotation_type ==\n                                      \"t_comparative\"\n                                        ? _c(\"span\", [\n                                            _vm._v(\n                                              \"\\n                      \" +\n                                                _vm._s(\n                                                  _vm.translations.show.brands\n                                                    .t_siemon_only\n                                                ) +\n                                                \"\\n                    \"\n                                            )\n                                          ])\n                                        : _c(\"span\", [\n                                            _vm._v(\n                                              \"\\n                      \" +\n                                                _vm._s(\n                                                  _vm.translations.show.brands[\n                                                    _vm.quotation.quotation_type\n                                                  ]\n                                                ) +\n                                                \"\\n                    \"\n                                            )\n                                          ])\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_footers\n                                            .not_used\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_footers\n                                            .not_used\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                      _vm._v(\n                                        \"\\n                    \" +\n                                          _vm._s(\n                                            _vm.currency.format(\n                                              _vm.quotation.products_totals\n                                                .t_siemon_only.without_percent\n                                            )\n                                          ) +\n                                          \"\\n                  \"\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-center\" }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_footers\n                                            .not_used\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                      _vm._v(\n                                        \"\\n                    \" +\n                                          _vm._s(\n                                            _vm.currency.format(\n                                              _vm.quotation.products_totals\n                                                .t_siemon_only.with_percent\n                                            )\n                                          ) +\n                                          \"\\n                  \"\n                                      )\n                                    ])\n                                  ],\n                                  1\n                                ),\n                                _vm._v(\" \"),\n                                _c(\n                                  \"b-tr\",\n                                  { staticClass: \"bg-dark text-center\" },\n                                  [\n                                    _c(\"b-td\", { attrs: { colspan: \"2\" } }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.translations.show.custom_headers\n                                            .total\n                                        )\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.currency.format(\n                                            parseFloat(\n                                              _vm.quotation.services_totals\n                                                .without_percent\n                                            ) +\n                                              parseFloat(\n                                                _vm.quotation.products_totals\n                                                  .t_siemon_only.without_percent\n                                              )\n                                          )\n                                        ) + \"\\n                  \"\n                                      )\n                                    ]),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\"),\n                                    _vm._v(\" \"),\n                                    _c(\"b-td\", { staticClass: \"text-right\" }, [\n                                      _vm._v(\n                                        _vm._s(\n                                          _vm.currency.format(\n                                            parseFloat(\n                                              _vm.quotation.services_totals\n                                                .with_percent\n                                            ) +\n                                              parseFloat(\n                                                _vm.quotation.products_totals\n                                                  .t_siemon_only.with_percent\n                                              )\n                                          )\n                                        ) + \"\\n                  \"\n                                      )\n                                    ])\n                                  ],\n                                  1\n                                )\n                              ]\n                            },\n                            proxy: true\n                          }\n                        ],\n                        null,\n                        false,\n                        4078951135\n                      )\n                    })\n                  : _vm._e()\n              ],\n              1\n            ),\n            _vm._v(\" \"),\n            _c(\"div\", { staticClass: \"col-12 mt-2\" }, [\n              _c(\"h5\", [\n                _c(\"b\", [_vm._v(_vm._s(_vm.translations.show.titles.credits))])\n              ]),\n              _vm._v(\" \"),\n              _c(\"span\", { staticClass: \"text-justify\" }, [\n                _vm._v(\n                  \"\\n              \" +\n                    _vm._s(_vm.quotation.credits) +\n                    \"\\n            \"\n                )\n              ])\n            ]),\n            _vm._v(\" \"),\n            _c(\"div\", { staticClass: \"col-12 mt-2\" }, [\n              _c(\"h5\", [\n                _c(\"b\", [_vm._v(_vm._s(_vm.translations.show.titles.warranty))])\n              ]),\n              _vm._v(\" \"),\n              _c(\"span\", { staticClass: \"text-justify\" }, [\n                _vm._v(\n                  \"\\n              \" +\n                    _vm._s(_vm.quotation.warranty) +\n                    \"\\n            \"\n                )\n              ])\n            ]),\n            _vm._v(\" \"),\n            _c(\"div\", { staticClass: \"col-12 mt-2\" }, [\n              _c(\"h5\", [\n                _c(\"b\", [\n                  _vm._v(_vm._s(_vm.translations.show.titles.payment_condition))\n                ])\n              ]),\n              _vm._v(\" \"),\n              _c(\"span\", { staticClass: \"text-justify\" }, [\n                _vm._v(\n                  \"\\n              \" +\n                    _vm._s(_vm.quotation.payment_condition) +\n                    \"\\n            \"\n                )\n              ])\n            ])\n          ])\n        ])\n      ])\n    ])\n  ])\n}\nvar staticRenderFns = []\nrender._withStripped = true\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/show.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return staticRenderFns; });\nvar render = function() {\n  var _vm = this\n  var _h = _vm.$createElement\n  var _c = _vm._self._c || _h\n  return _c(\"div\", { staticClass: \"row\" }, [\n    _c(\n      \"div\",\n      { staticClass: \"col-lg-12 offset-xl-1 col-xl-10\" },\n      [\n        _c(\"global-view\", {\n          attrs: { quotation_id: parseInt(this.$route.params.id) }\n        })\n      ],\n      1\n    )\n  ])\n}\nvar staticRenderFns = []\nrender._withStripped = true\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/show.vue?./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options");
 
 /***/ }),
 
@@ -3628,6 +3652,42 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _nod
 
 "use strict";
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_conditions_vue_vue_type_template_id_cfacd9a2___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../node_modules/vue-loader/lib??vue-loader-options!./conditions.vue?vue&type=template&id=cfacd9a2& */ \"./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/components/conditions.vue?vue&type=template&id=cfacd9a2&\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_conditions_vue_vue_type_template_id_cfacd9a2___WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_conditions_vue_vue_type_template_id_cfacd9a2___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"]; });\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/conditions.vue?");
+
+/***/ }),
+
+/***/ "./viewsrc/modules/quotations/components/global_view.vue":
+/*!***************************************************************!*\
+  !*** ./viewsrc/modules/quotations/components/global_view.vue ***!
+  \***************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _global_view_vue_vue_type_template_id_96eec44c___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./global_view.vue?vue&type=template&id=96eec44c& */ \"./viewsrc/modules/quotations/components/global_view.vue?vue&type=template&id=96eec44c&\");\n/* harmony import */ var _global_view_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./global_view.vue?vue&type=script&lang=js& */ \"./viewsrc/modules/quotations/components/global_view.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ \"./node_modules/vue-loader/lib/runtime/componentNormalizer.js\");\n\n\n\n\n\n/* normalize component */\n\nvar component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"])(\n  _global_view_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__[\"default\"],\n  _global_view_vue_vue_type_template_id_96eec44c___WEBPACK_IMPORTED_MODULE_0__[\"render\"],\n  _global_view_vue_vue_type_template_id_96eec44c___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"],\n  false,\n  null,\n  null,\n  null\n  \n)\n\n/* hot reload */\nif (false) { var api; }\ncomponent.options.__file = \"viewsrc/modules/quotations/components/global_view.vue\"\n/* harmony default export */ __webpack_exports__[\"default\"] = (component.exports);\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/global_view.vue?");
+
+/***/ }),
+
+/***/ "./viewsrc/modules/quotations/components/global_view.vue?vue&type=script&lang=js&":
+/*!****************************************************************************************!*\
+  !*** ./viewsrc/modules/quotations/components/global_view.vue?vue&type=script&lang=js& ***!
+  \****************************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_index_js_vue_loader_options_global_view_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/vue-loader/lib??vue-loader-options!./global_view.vue?vue&type=script&lang=js& */ \"./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/components/global_view.vue?vue&type=script&lang=js&\");\n/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__[\"default\"] = (_node_modules_vue_loader_lib_index_js_vue_loader_options_global_view_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__[\"default\"]); \n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/global_view.vue?");
+
+/***/ }),
+
+/***/ "./viewsrc/modules/quotations/components/global_view.vue?vue&type=template&id=96eec44c&":
+/*!**********************************************************************************************!*\
+  !*** ./viewsrc/modules/quotations/components/global_view.vue?vue&type=template&id=96eec44c& ***!
+  \**********************************************************************************************/
+/*! exports provided: render, staticRenderFns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_global_view_vue_vue_type_template_id_96eec44c___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../node_modules/vue-loader/lib??vue-loader-options!./global_view.vue?vue&type=template&id=96eec44c& */ \"./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./viewsrc/modules/quotations/components/global_view.vue?vue&type=template&id=96eec44c&\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_global_view_vue_vue_type_template_id_96eec44c___WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"staticRenderFns\", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_global_view_vue_vue_type_template_id_96eec44c___WEBPACK_IMPORTED_MODULE_0__[\"staticRenderFns\"]; });\n\n\n\n//# sourceURL=webpack:///./viewsrc/modules/quotations/components/global_view.vue?");
 
 /***/ }),
